+++ src/helper/formatModifier.js
@@ -1,0 +1,8 @@
+export function formatModifier(number) {
+    switch (true) {
+        case (isNaN(number)): return '-';
+        case (number > 0): return `+${number}`;
+        case (number === 0): return `+${number}`;
+        case (number < 0): return `-${number}`;
+    }
+}